pyproject.toml
[tool.poetry]
name = "eigenface"
version = "0.1.0"
description = "Eigenface face recognition project for University of Helsinki's Data structures and algorithms course."
authors = ["ni-eminen <matiasnie2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.23.4"
pillow = "^9.3.0"
pytest = "^7.2.0"
coverage = "^6.5.0"
ipynb = "^0.5.1"
jupyter = "^1.0.0"
sklearn = "^0.0.post1"
scikit-learn = "^1.1.3"
pandoc = "^2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
